requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
deskew==1.3.3
distro==1.8.0
easyocr==1.6.2
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
imageio==2.25.0
imgaug==0.4.0
jamspell==0.0.12
joblib==1.2.0
keras==2.11.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
ninja==1.11.1
nltk==3.8.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pyparsing==3.0.9
pyspellchecker==0.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-build==0.16.6
scikit-image==0.19.3
scipy==1.10.0
shapely==2.0.0
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
textblob==0.17.1
tifffile==2023.1.23.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
validators==0.20.0
Werkzeug==2.2.2
wrapt==1.14.1
